Move all exceptions to exceptions.py
