+++ pymc/__init__.py
@@ -47,12 +47,11 @@ def __set_compiler_flags():
 
 __set_compiler_flags()
 
-from pymc import _version, gp, math, ode, plots, sampling, stats
+from pymc import _version, exceptions, gp, math, ode, plots, sampling, stats
 from pymc.backends import *
 from pymc.blocking import *
 from pymc.data import *
 from pymc.distributions import *
-from pymc.exceptions import *
 from pymc.func_utils import find_constrained_prior
 from pymc.logprob import *
 from pymc.math import (
